+++ packages/manager/src/features/LoadBalancers/LoadBalancerDetail/LoadBalancerConfigurations.tsx
@@ -1,15 +1,22 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { Waypoint } from 'react-waypoint';
 
+import { Box } from 'src/components/Box';
+import { Button } from 'src/components/Button/Button';
 import { CircleProgress } from 'src/components/CircleProgress';
 import { ErrorState } from 'src/components/ErrorState/ErrorState';
+import { Paper } from 'src/components/Paper';
+import { Stack } from 'src/components/Stack';
 import { useLoabalancerConfigurationsInfiniteQuery } from 'src/queries/aglb/configurations';
 
 import { ConfigurationAccordion } from './Configurations/ConfigurationAccordion';
+import { ConfigurationForm } from './Configurations/ConfigurationForm';
 
 export const LoadBalancerConfigurations = () => {
   const { loadbalancerId } = useParams<{ loadbalancerId: string }>();
+
+  const [isCreating, setIsCreating] = useState(false);
 
   const {
     data,
@@ -36,11 +43,25 @@ export const LoadBalancerConfigurations = () => {
         <ConfigurationAccordion
           configuration={configuration}
           key={configuration.id}
-          loadbalancerId={Number(loadbalancerId)}
         />
       ))}
       {hasNextPage && <Waypoint onEnter={() => fetchNextPage()} />}
       {isFetchingNextPage && <CircleProgress mini />}
+      <Stack my={2} spacing={2}>
+        <Box>
+          <Button buttonType="outlined" onClick={() => setIsCreating(true)}>
+            Add {configurations?.length === 0 ? '' : 'another'} Configuration
+          </Button>
+        </Box>
+        {isCreating && (
+          <Paper>
+            <ConfigurationForm
+              mode="create"
+              onCancel={() => setIsCreating(false)}
+            />
+          </Paper>
+        )}
+      </Stack>
     </>
   );
 };
