+++ packages/manager/src/mocks/serverHandlers.ts
@@ -357,7 +357,11 @@ const aglb = [
   rest.get('*/v4beta/aglb/:id/routes', (req, res, ctx) => {
     const headers = JSON.parse(req.headers.get('x-filter') || '{}');
     if (headers['+or']) {
-      return res(ctx.json(makeResourcePage(routeFactory.buildList(2))));
+      return res(
+        ctx.json(
+          makeResourcePage(routeFactory.buildList(headers['+or'].length))
+        )
+      );
     }
     return res(ctx.json(makeResourcePage(routeFactory.buildList(5))));
   }),
